+++ OneBlink.SDK/models/GetSubmissionPdfRequest.cs
@@ -25,6 +25,10 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public bool? includeCalendarBookingInPdf
+        {
+            get; set;
+        }
         public List<string> excludedCSSClasses
         {
             get; set;